refactor(access): migrate sub-model viewset logic to common
ref: #724 #723
