+++ app/api/viewsets/common.py
@@ -727,6 +727,115 @@ class ModelViewSet(
 
 
 
+class SubModelViewSet(
+    ModelViewSetBase,
+    Create,
+    Retrieve,
+    Update,
+    Destroy,
+    List,
+    viewsets.ModelViewSet,
+):
+
+    base_model = None
+    """Model that is the base of this sub-model"""
+
+    model_kwarg: str = None
+    """Kwarg name for the sub-model"""
+
+
+    @property
+    def model(self):
+
+
+        if getattr(self, '_model', None) is not None:
+
+            return self._model
+
+        model_kwarg = None
+
+        if hasattr(self, 'kwargs'):
+
+            model_kwarg = self.kwargs.get(self.model_kwarg, None)
+
+        if model_kwarg:
+
+            self._model = self.related_objects(self.base_model, model_kwarg)
+
+        else:
+
+            self._model = self.base_model
+
+        return self._model
+
+
+    def related_objects(self, model, model_kwarg):
+        """Recursive relate_objects fetch
+
+        Fetch the model that is lowest in the chain of inherited models
+
+        Args:
+            model (django.db.models.Model): The model to obtain the 
+                related_model from.
+            model_kwarg (str): The URL Kwarg of the model.
+
+        Returns:
+            _type_: _description_
+        """
+
+        related_model = None
+
+        if model_kwarg:
+
+            for related_object in model._meta.related_objects:
+
+                related_objects = getattr(related_object.related_model._meta, 'related_objects', [])
+
+                if(
+                    str(
+                        related_object.related_model._meta.verbose_name
+                    ).lower().replace(' ', '_') == model_kwarg
+                ):
+
+                    related_model = related_object.related_model
+                    break
+                
+                elif related_objects:
+
+                    related_model = self.related_objects(model = related_object.related_model, model_kwarg = model_kwarg)
+
+                    break
+
+
+        return related_model
+
+    def get_serializer_class(self):
+
+        serializer_module = importlib.import_module(
+            self.model._meta.app_label + '.serializers.' + str(
+                self.model._meta.verbose_name
+            ).lower().replace(' ', '_')
+        )
+
+        if (
+            self.action == 'list'
+            or self.action == 'retrieve'
+        ):
+
+            self.serializer_class = getattr(serializer_module, 'ViewSerializer')
+
+
+        else:
+
+            self.serializer_class = getattr(serializer_module, 'ModelSerializer')
+
+
+        return self.serializer_class
+
+
+
+
+
 class ModelCreateViewSet(
     ModelViewSetBase,
     Create,
